Working on Login page and Establishment configuration

+++ app/Http/Livewire/DossiersTable.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Laracasts\Flash\Flash;
 use Rappasoft\LaravelLivewireTables\DataTableComponent;
 use Rappasoft\LaravelLivewireTables\Views\Column;
@@ -9,12 +10,15 @@ use App\Models\Dossier;
 
 class DossiersTable extends DataTableComponent
 {
+    use AuthorizesRequests;
     protected $model = Dossier::class;
 
     protected $listeners = ['deleteRecord' => 'deleteRecord'];
 
     public function deleteRecord($id)
     {
+        $this->authorize('delete', Dossier::class);
+
         Dossier::find($id)->delete();
         Flash::success(__('messages.deleted', ['model' => __('models/dossiers.singular')]));
         $this->emit('refreshDatatable');
