+++ app/Http/Livewire/UsersTable.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Laracasts\Flash\Flash;
 use Rappasoft\LaravelLivewireTables\DataTableComponent;
 use Rappasoft\LaravelLivewireTables\Views\Column;
@@ -9,12 +10,14 @@ use App\Models\User;
 
 class UsersTable extends DataTableComponent
 {
+    use AuthorizesRequests;
     protected $model = User::class;
 
     protected $listeners = ['deleteRecord' => 'deleteRecord'];
 
     public function deleteRecord($id)
     {
+        $this->authorize('delete', User::class);
         User::find($id)->delete();
         Flash::success(__('messages.deleted', ['model' => __('models/users.singular')]));
         $this->emit('refreshDatatable');
